# contracts/platform/api/openapi.yaml
openapi: "3.1.0"
...
info:
  title: BlueLab v1 API
  version: 1.0.0
  summary: Voice-roleplay training and hiring platform — product, candidate, and operations surfaces.
  description: >
    Contract-first specification for the BlueLab application plane (F-1). Three principals:
    account holders (manager, rep) via session cookie; candidates via bearer capability token;
    BlueLab Internal Operations via a separate ops session. Cross-scope denial is an
    indistinguishable 404 (ADR-0036). All errors are RFC 9457 application/problem+json.
    Timestamps are UTC RFC 3339; months are YYYY-MM interpreted in the org's timezone;
    scores are 0-10 with one decimal and server-computed tier bands.
  license:
    name: Proprietary — BlueLab. All rights reserved.
    url: https://bluelab.example/license
  contact:
    name: BlueLab Engineering
servers:
  - url: https://api.bluelab.example
    description: Placeholder origin — Phase 7 owns real hosts. Paths carry their own /api/v1, /ops/v1, /hooks prefixes.
  - url: http://127.0.0.1:4010
    description: Prism mock server (contract walkthroughs).
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/positions:
    get:
      tags: [Hiring]
      summary: Positions list
      description: >
        The manager's positions with status (needs authoring / active / closed) and the pipeline
        counts — invited, pending review, approved-unsent — that drive the adaptive primary action
        and the summary cards (FR-HIR-001/002).
      operationId: listPositions
      parameters:
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Positions page.
          content:
            application/json:
              schema:
                type: object
                required: [data, totals, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/PositionSummary" }
                  totals:
                    type: object
                    description: Server-computed aggregates across ALL the manager's positions, independent of pagination (FR-HIR-002; gate F-4).
                    required: [open_positions, pending_review, approved_unsent]
                    properties:
                      open_positions: { type: integer, minimum: 0 }
                      pending_review: { type: integer, minimum: 0 }
                      approved_unsent: { type: integer, minimum: 0 }
                  pagination: { $ref: "#/components/schemas/CursorPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
    post:
      tags: [Hiring]
      summary: Create a position
      description: >
        Creates a position (FR-HIR-003): candidate-facing title, openings, completion-notification
        toggle (E-5), candidate-report policy (default rejected_only) governing E-3 delivery
        (FR-CND-012). Call language displays as Egyptian Arabic, fixed.
      operationId: createPosition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PositionCreate" }
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "201":
          description: Created (status needs_authoring).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PositionDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/positions/{position_id}:
    get:
      tags: [Hiring]
      summary: Position detail
      description: Settings, assessment stages in order, freeze state, and pipeline counts. Closed positions are readable archives (FR-HIR-015).
      operationId: getPosition
      parameters:
        - $ref: "#/components/parameters/PositionIdParam"
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Position.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PositionDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
    patch:
      tags: [Hiring]
      summary: Update position settings
      description: >
        Edits settings — title, openings, notification toggle, report policy, invite expiry preset,
        invite email template (FR-HIR-003/008). The assessment itself changes only via PUT
        /assessment; closed positions answer 409 position-closed.
      operationId: updatePosition
      parameters:
        - $ref: "#/components/parameters/PositionIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PositionPatch" }
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PositionDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem type: position-closed."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/positions/{position_id}/assessment:
    put:
      tags: [Hiring]
      summary: Compose the assessment (ordered drill set)
      description: >
        Full-replace of the ordered stage list from the team's published catalog (FR-HIR-004); any
        count, freely reordered; inline-authored drills join the catalog via the standard flow and
        are then referenced here. Saving with at least one drill activates the position
        (FR-HIR-005). After the first invite the set and order are frozen: 409 assessment-frozen
        (AC-HIR-001). Every candidate faces the identical assessment (CMP-003).
      operationId: putAssessment
      parameters:
        - $ref: "#/components/parameters/PositionIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [drill_ids]
              properties:
                drill_ids:
                  type: array
                  minItems: 1
                  items: { type: string, format: uuid }
                  description: Published team drills, in stage order.
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Assessment saved; position active.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PositionDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem types: assessment-frozen, position-closed, drill-not-startable (unpublished/archived drill in the set)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/positions/{position_id}/candidates:
    get:
      tags: [Hiring]
      summary: Pipeline views
      description: >
        The position's candidates in three views with live counts (FR-HIR-010): pipeline (everyone,
        journey state + decision state + latest invite delivery), review_queue (completed, ranked
        by overall score descending), approved (with contact details). Incomplete candidates
        (expired mid-assessment or consumed stage) are marked and remain decidable on partial
        evidence (FR-HIR-016).
      operationId: listPositionCandidates
      parameters:
        - $ref: "#/components/parameters/PositionIdParam"
        - name: view
          in: query
          schema: { type: string, enum: [pipeline, review_queue, approved], default: pipeline }
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Candidates page with tab counts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PipelinePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
    post:
      tags: [Hiring]
      summary: Add a candidate batch
      description: >
        Adds candidates as entries of name, email, phone, LinkedIn, source, plus the
        manager-private internal note (FR-HIR-006 — never visible to any candidate). Batch is
        atomic. Requires Idempotency-Key (ADR-0040). Closed positions answer 409.
      operationId: addCandidates
      parameters:
        - $ref: "#/components/parameters/PositionIdParam"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [candidates]
              properties:
                candidates:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: { $ref: "#/components/schemas/CandidateCreate" }
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "201":
          description: Candidates created (not yet invited).
          content:
            application/json:
              schema:
                type: object
                required: [created]
                properties:
                  created:
                    type: array
                    items:
                      type: object
                      required: [candidate_id, name, email]
                      properties:
                        candidate_id: { type: string, format: uuid }
                        name: { type: string }
                        email: { type: string, format: email }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem types: position-closed, idempotency-key-reuse."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          description: "Problem types: validation-error; idempotency-key-required when the header is missing (gate F-5)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ValidationProblem" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/positions/{position_id}/candidates/parse:
    post:
      tags: [Hiring]
      summary: Parse a bulk candidate file
      description: >
        Parses an uploaded list in the downloadable template format into rows that pre-fill the
        batch for review and editing (FR-HIR-007). Never creates candidates or sends invites —
        sending always happens from the reviewed batch via the endpoints above. The downloadable
        template itself is an SPA-bundled static asset versioned with the frontend, not an API
        resource (gate F-9).
      operationId: parseCandidateFile
      parameters:
        - $ref: "#/components/parameters/PositionIdParam"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Parsed rows with per-row problems for review.
          content:
            application/json:
              schema:
                type: object
                required: [rows]
                properties:
                  rows:
                    type: array
                    items: { $ref: "#/components/schemas/ParsedCandidateRow" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "415":
          description: "Problem type: unsupported-file-type (template mismatch)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/positions/{position_id}/invites:
    post:
      tags: [Hiring]
      summary: Send invites
      description: >
        The T-7 transaction (FR-HIR-008): freezes the assessment on first send (FR-HIR-005), issues
        each candidate a unique token bound to org/position/candidate (FR-IDA-011) with the
        position's expiry, and queues each invite email (E-2) with the token link in the URL
        fragment. One attempt per drill per the assessment rules. Requires Idempotency-Key
        (ADR-0040); email dedupe per token backstops at-most-once sends.
      operationId: sendInvites
      parameters:
        - $ref: "#/components/parameters/PositionIdParam"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [candidate_ids]
              properties:
                candidate_ids:
                  type: array
                  minItems: 1
                  items: { type: string, format: uuid }
                expiry_days:
                  type: integer
                  minimum: 1
                  maximum: 60
                  description: Overrides the position preset for this send (FR-HIR-008; default 7).
                invite_template:
                  type: [string, "null"]
                  description: Edited email body for this send; null uses the position template.
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "202":
          description: Tokens issued; emails queued; assessment frozen from the first send.
          content:
            application/json:
              schema:
                type: object
                required: [invites_queued, assessment_frozen]
                properties:
                  invites_queued: { type: integer, minimum: 1 }
                  assessment_frozen: { type: boolean }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem types: assessment-empty, position-closed, idempotency-key-reuse."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          description: "Problem types: validation-error; idempotency-key-required when the header is missing (gate F-5)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ValidationProblem" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/candidates/{candidate_id}:
    patch:
      tags: [Hiring]
      summary: Update a candidate's internal note
      description: >
        Edits the manager-private internal note (FR-HIR-006) that resurfaces on the report
        (FR-HIR-011). Never reaches anything a candidate can see; excluded from the PDF by
        construction (FR-HIR-012).
      operationId: updateCandidate
      parameters:
        - $ref: "#/components/parameters/CandidateIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [internal_note]
              properties:
                internal_note: { type: [string, "null"], maxLength: 4000 }
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "200":
          description: Updated pipeline row.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PipelineCandidate" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem type: position-closed."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/candidates/{candidate_id}/invite-resend:
    post:
      tags: [Hiring]
      summary: Resend an invite
      description: >
        Re-sends the templated invite with a fresh token and reset expiry; the prior token stays
        valid until its own original expiry (FR-HIR-009, FR-IDA-012, AC-IDA-005). Confirmation is
        client-side. A resend after expiry-mid-assessment resumes the same assessment state
        (FR-CND-011). Throttled per 05-rate-limits §5.
      operationId: resendInvite
      parameters:
        - $ref: "#/components/parameters/CandidateIdParam"
      responses:
        "403": { $ref: "#/components/responses/OrgSuspended" }
        "202":
          description: Fresh token issued; email queued.
          content:
            application/json:
              schema:
                type: object
                required: [expires_at]
                properties:
                  expires_at: { type: string, format: date-time }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Problem types: position-closed, assessment-completed (nothing to resend for a completed candidate)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  parameters:
    CursorParam:
      name: cursor
      in: query
      description: Opaque pagination cursor from a previous response.
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
    PositionIdParam:
      name: position_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-bluelab_session
      description: Account-holder server-side session (ADR-0037).
    candidateToken:
      type: http
      scheme: bearer
      description: Candidate capability token from the invite link fragment (FR-IDA-011, ADR-0037).
    opsSessionCookie:
      type: apiKey
      in: cookie
      name: __Host-bluelab_ops_session
      description: BlueLab Internal Operations session — separate surface (ADR-0010).
    livekitWebhook:
      type: apiKey
      in: header
      name: Authorization
      description: LiveKit-signed JWT carrying a SHA-256 of the payload (01-realtime-call-contract §7).